Move internal ConfigRepoOperationsControllerV1 endpoints into ConfigReposInternalControllerV1
  - Remove `last_parsed_result` endpoint is it was not used and is redundant with the internal config_repos GET endpoint
  - The ConfigRepoOperationsControllerV1 will be repurposed as a public API for `preflight` checks
