+++ api/api-config-repo-operations-v1/src/test/groovy/com/thoughtworks/go/apiv1/configrepooperations/ConfigRepoOperationsControllerV1Test.groovy
@@ -60,196 +60,4 @@ class ConfigRepoOperationsControllerV1Test implements SecurityServiceTrait, Cont
   ConfigRepoOperationsControllerV1 createControllerInstance() {
     return new ConfigRepoOperationsControllerV1(new ApiAuthenticationHelper(securityService, goConfigService), dataSource, service, materialUpdateService, converter)
   }
-
-  @Nested
-  class Security implements SecurityTestTrait, AdminUserSecurity {
-    @Override
-    String getControllerMethodUnderTest() {
-      return "getLastParseResult"
-    }
-
-    @Override
-    void makeHttpCall() {
-      getWithApiHeader(controller.controllerPath(ID, 'last_parsed_result'), [:])
-    }
-  }
-
-  @Nested
-  class getLastParsedResult {
-    @BeforeEach
-    void setUp() {
-      loginAsAdmin()
-    }
-
-    @Test
-    void 'should get last parsed result for config repo'() {
-      MaterialConfig material = mock(MaterialConfig.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(material, null, ID))
-      when(dataSource.getLastParseResult(material)).thenReturn(new PartialConfigParseResult("1234", new PartialConfig()))
-
-      getWithApiHeader(controller.controllerPath(ID, 'last_parsed_result'), [:])
-
-      assertThatResponse()
-        .isOk()
-        .hasContentType(controller.mimeType)
-        .hasJsonBody([revision: "1234", success: true])
-    }
-
-    @Test
-    void 'should see any errors from last parsed result'() {
-      MaterialConfig material = mock(MaterialConfig.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(material, null, ID))
-      when(dataSource.getLastParseResult(material)).thenReturn(new PartialConfigParseResult("1234", new RuntimeException("Boom!")))
-
-      getWithApiHeader(controller.controllerPath(ID, 'last_parsed_result'), [:])
-
-      assertThatResponse()
-        .isOk()
-        .hasContentType(controller.mimeType)
-        .hasJsonBody([revision: "1234", success: false, error: "Boom!"])
-    }
-
-    @Test
-    void 'should handle materials that have never been seen'() {
-      MaterialConfig material = mock(MaterialConfig.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(material, null, ID))
-      when(dataSource.getLastParseResult(material)).thenReturn(null)
-
-      getWithApiHeader(controller.controllerPath(ID, 'last_parsed_result'), [:])
-
-      assertThatResponse()
-        .isOk()
-        .hasContentType(controller.mimeType)
-        .hasJsonBody([revision: null])
-    }
-
-    @Test
-    void 'should handle non-existent config repo'() {
-      when(service.getConfigRepo(ID)).thenReturn(null)
-
-      getWithApiHeader(controller.controllerPath(ID, 'last_parsed_result'), [:])
-
-      assertThatResponse()
-        .isNotFound()
-        .hasContentType(controller.mimeType)
-        .hasJsonMessage(HaltApiMessages.notFoundMessage())
-    }
-  }
-
-  @Nested
-  class StatusSecurity implements SecurityTestTrait, AdminUserSecurity {
-    @Override
-    String getControllerMethodUnderTest() {
-      return "inProgress"
-    }
-
-    @Override
-    void makeHttpCall() {
-      getWithApiHeader(controller.controllerPath(ID, 'status'), [:])
-    }
-  }
-
-  @Nested
-  class status {
-    @BeforeEach
-    void setUp() {
-      loginAsAdmin()
-    }
-
-    @Test
-    void 'should confirm if update is in progress'() {
-      MaterialConfig config = mock(MaterialConfig.class)
-      Material material = mock(Material.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(config, null, ID))
-      when(converter.toMaterial(config)).thenReturn(material)
-      when(materialUpdateService.isInProgress(material)).thenReturn(true)
-
-      getWithApiHeader(controller.controllerPath(ID, 'status'), [:])
-      verify(materialUpdateService).isInProgress(material)
-
-      assertThatResponse()
-        .isOk()
-        .hasContentType(controller.mimeType)
-        .hasJsonBody([inProgress: true])
-    }
-
-    @Test
-    void 'should return false if update is not in progress'() {
-      MaterialConfig config = mock(MaterialConfig.class)
-      Material material = mock(Material.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(config, null, ID))
-      when(converter.toMaterial(config)).thenReturn(material)
-      when(materialUpdateService.isInProgress(material)).thenReturn(false)
-
-      getWithApiHeader(controller.controllerPath(ID, 'status'), [:])
-      verify(materialUpdateService).isInProgress(material)
-
-      assertThatResponse()
-        .isOk()
-        .hasContentType(controller.mimeType)
-        .hasJsonBody([inProgress: false])
-    }
-  }
-
-  @Nested
-  class TriggerSecurity implements SecurityTestTrait, AdminUserSecurity {
-    @Override
-    String getControllerMethodUnderTest() {
-      return "triggerUpdate"
-    }
-
-    @Override
-    void makeHttpCall() {
-      postWithApiHeader(controller.controllerPath(ID, 'trigger_update'), [:])
-    }
-  }
-
-  @Nested
-  class triggerUpdate {
-    @BeforeEach
-    void setUp() {
-      loginAsAdmin()
-    }
-
-    @Test
-    void 'should trigger material update for repo'() {
-      MaterialConfig config = mock(MaterialConfig.class)
-      Material material = mock(Material.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(config, null, ID))
-      when(converter.toMaterial(config)).thenReturn(material)
-      when(materialUpdateService.updateMaterial(material)).thenReturn(true)
-
-      postWithApiHeader(controller.controllerPath(ID, 'trigger_update'), [:])
-      verify(materialUpdateService).updateMaterial(material)
-
-      assertThatResponse()
-        .isCreated()
-        .hasContentType(controller.mimeType)
-        .hasJsonMessage("OK")
-    }
-
-    @Test
-    void 'should not trigger update if update is already in progress'() {
-      MaterialConfig config = mock(MaterialConfig.class)
-      Material material = mock(Material.class)
-
-      when(service.getConfigRepo(ID)).thenReturn(new ConfigRepoConfig(config, null, ID))
-      when(converter.toMaterial(config)).thenReturn(material)
-      when(materialUpdateService.updateMaterial(material)).thenReturn(false)
-
-      postWithApiHeader(controller.controllerPath(ID, 'trigger_update'), [:])
-      verify(materialUpdateService).updateMaterial(material)
-
-      assertThatResponse()
-        .isConflict()
-        .hasContentType(controller.mimeType)
-        .hasJsonMessage("Update already in progress.")
-    }
-  }
 }
