+++ server/webapp/WEB-INF/urlrewrite.xml
@@ -250,7 +250,7 @@
   <rule>
     <name>Spark Config Repos Internal Operations API</name>
     <condition type="header" name="Accept">application\/vnd\.go\.cd\.v1\+json</condition>
-    <from>^/api/internal/config_repos/([^/]+)/(last_parsed_result|trigger_update|status)(/?)</from>
+    <from>^/api/internal/config_repos/([^/]+)/(trigger_update|status)(/?)</from>
     <to last="true">/spark/api/internal/config_repos/${escape:$1}/${escape:$2}</to>
   </rule>
